% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsim_th_tau2.R
\name{dsim_th_tau2}
\alias{dsim_th_tau2}
\title{Simultaneous distribution of overall treatment effect and between-study variance}
\usage{
dsim_th_tau2(y, x, theta = 0, tau2, K, vi, method = "DL", n_ftau2 = 1e+05)
}
\arguments{
\item{theta}{value of true overall effect that define 0 in default.}

\item{tau2}{between-study variance.}

\item{K}{number of studies.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{n_ftau2}{value of number of sampling for estimates of between-study variance. The default is 1000.}
}
\value{
value of simultaneous distribution of overall treatment effect and between-study variance.
}
\description{
Simultaneous distribution of overall treatment effect and between-study variance
}
