% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsim_th_tau2.R
\name{dtau2}
\alias{dtau2}
\title{Density function of between-study variance}
\usage{
dtau2(x, tau2, K, vi, method = "DL", n_ftau2 = 1e+05)
}
\arguments{
\item{x}{value of between-study variance.}

\item{tau2}{between-study variance.}

\item{K}{number of studies.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{n_ftau2}{value of number of sampling for estimates of between-study variance. The default is 100000.}

\item{obj}{character string to specify the objective. "Estimate", which set to default, is overall treatment effect. "Predict" is predictive treatment effect.}
}
\value{
value of simultaneous distribution of overall treatment effect and between-study variance.
}
\description{
Density function of between-study variance
}
