% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_tau2m.R
\name{est_tau2m}
\alias{est_tau2m}
\title{Simultaneous distribution of overall treatment effect and between-study variance}
\usage{
est_tau2m(
  yi,
  vi,
  K,
  method = "DL",
  rho = "ML",
  n_ftau2,
  tau2.min = 1e-10,
  tau2.max = 100,
  seed = 1234
)
}
\arguments{
\item{yi}{vector of length K with the observed effect sizes or outcomes.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{K}{number of studies.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{rho}{the character to specify the convex function. By default (rho="ML"), maximum likelihood is used. When rho="L1", absolute function is used. When rho="L2", square function is used.}

\item{n_ftau2}{value of number of sampling for estimates of between-study variance. The default is 1000.}

\item{tau2.min}{value to be added so that the logarithm does not go to infinity. By default, 1e-10 is used.}

\item{tau2.max}{maximum search range in minimizing the estimation equation for M-estimation. The default is 100, but a smaller value may be appropriate depending on the between-study variance.}

\item{seed}{random number seed used within the function. By default, 1234 is used.}
}
\value{
M-estimator of between-study variance by a specific method.
}
\description{
Simultaneous distribution of overall treatment effect and between-study variance
}
