% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun}
\alias{estfun}
\title{Estimation function of between-study variance}
\usage{
estfun(tau2m, tau2h, K, vi, method = "DL", n_ftau2 = 1000, func = 1, seed = 1)
}
\arguments{
\item{tau2m}{M-estimator of between-study variance.}

\item{tau2h}{estimator of between-study variance by setting method. The default is "DL" method.}

\item{K}{number of studies.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{n_ftau2}{value of number of sampling for estimates of between-study variance. The default is 1000.}

\item{func}{If 1 (default), the convex function is absolute function. If 2, the convex function is square function.}

\item{seed}{seed value of random numbers.}
}
\value{
value of estimation function for between-study variance.
}
\description{
Estimation function of between-study variance
}
