% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMest-package.R
\docType{package}
\name{metaMest-package}
\alias{metaMest-package}
\alias{metaMest}
\title{The 'metaMest' package.}
\description{
Meta-analysis by M-estimator based approach described in Hanada and Sugimoto (2024)
}
\references{
Hanada, K., & Sugimoto, T. (2024). Random-Effect Meta-Analysis with Robust Between-Study Variance. Pre-print.
}
\author{
\strong{Maintainer}: Keisuke Hanada \email{keisuke.hanada.87@gmail.com} (\href{https://orcid.org/0000-0002-1444-0280}{ORCID})

}
