% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dsim_th_tau2.R
\name{plot_dsim_th_tau2}
\alias{plot_dsim_th_tau2}
\title{Plot function for simultaneous distribution of overall treatment effect and between-study variance}
\usage{
plot_dsim_th_tau2(
  contour_val = c(0.01, 0.05, 0.2, 0.4, 0.6, 0.8),
  theta = 0,
  tau2,
  K,
  vi,
  alpha = 0.05,
  theta.limit = c(-10, 10),
  tau2.limit = c(0, 10),
  length.out = 1000,
  n = 1e+05,
  method = "DL",
  obj = "Estimate"
)
}
\arguments{
\item{contour_val}{vector of the thresholds for contour plot. By default, the line of (0.01, 0.05, 0.2, 0.4, 0.6, 0.8) is plotted.}

\item{theta}{value of true overall effect that define 0 in default.}

\item{tau2}{between-study variance.}

\item{K}{number of studies.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{alpha}{significance level of simultaneous plot.}

\item{theta.limit}{lower and upper limits for plotting theta.}

\item{tau2.limit}{lower and upper limits for plotting tau2.}

\item{length.out}{x-axis and y-axis step widths}

\item{n}{number of sampling for calculating the thresholds. The default is 100000.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{obj}{character string to specify the objective. "Estimate", which set to default, is overall treatment effect. "Predict" is predictive treatment effect.}
}
\value{
contour plot of simultaneous distribution of overall treatment effect and between-study variance.
}
\description{
Plot function for simultaneous distribution of overall treatment effect and between-study variance
}
