% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-tau2DL.R
\name{rand_tau2dl_tau2}
\alias{rand_tau2dl_tau2}
\title{Sampling from exact distribution of DerSimonian-Laird between-study variance estimator}
\usage{
rand_tau2dl_tau2(n, K, tau2, sig2k)
}
\arguments{
\item{n}{number of sampling}

\item{K}{number of studies}

\item{tau2}{between-study variance parameter}

\item{sig2k}{within-study variance in each study}
}
\value{
sampling values from exact distribution of DerSimonian-Laird between-study variance estimator
}
\description{
Sampling from exact distribution of DerSimonian-Laird between-study variance estimator
}
