% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Convex function for estimation of M-estimator}
\usage{
rho(u, func = 1)
}
\arguments{
\item{u}{Value of variable.}

\item{func}{If 1 (default), the convex function is absolute function. If 2, the convex function is square function.}
}
\value{
value of convex function
}
\description{
Convex function for estimation of M-estimator
}
