% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-Muni.R
\name{rma.Muni.m1}
\alias{rma.Muni.m1}
\title{Random-effects meta-analysis with M-estimator}
\usage{
rma.Muni.m1(
  yi,
  vi,
  theta,
  tau2,
  alpha = 0.05,
  method = "DL",
  interval = c(-10, 10),
  h = 1000
)
}
\arguments{
\item{yi}{vector of length K with the observed effect sizes or outcomes.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{theta}{value of estimator for the overall treatment effect.}

\item{tau2}{value of estimator for the between-study variance.}

\item{alpha}{significance level of confidence intervals, and simultaneous plot.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{interval}{search range of confidence interval for overall treatment effect.}

\item{h}{search grid of confidence interval for overall treatment effect.}
}
\value{
confidence interval for M-estimated random-effects meta-analysis with absolute convex function.
}
\description{
Random-effects meta-analysis with M-estimator
}
