% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma-Muni.R
\name{rma.Muni}
\alias{rma.Muni}
\title{Random-effects meta-analysis with M-estimator}
\usage{
rma.Muni(
  yi,
  vi,
  method = "DL",
  HKSJ = FALSE,
  alpha = 0.05,
  n.rand = 10000,
  tau2.max = 100,
  epsilon = 1e-10,
  theta.min = -100,
  theta.max = 100,
  n.grid = 300,
  obj = "Estimate",
  contour_val = c(0.01, 0.05, 0.2, 0.4, 0.6, 0.8),
  plot = TRUE
)
}
\arguments{
\item{yi}{vector of length K with the observed effect sizes or outcomes.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}

\item{HKSJ}{If TRUE, HKSJ approach is applied. HKSJ=FALSE is default.}

\item{alpha}{significance level of confidence intervals, and simultaneous plot.}

\item{n.rand}{sampling number of random values for distribution of between-study variance.}

\item{tau2.max}{upper bound for M-estimate of tau2.}

\item{theta.min}{lower bound for M-estimate of theta.}

\item{n.grid}{number of grid for contour plot of treatment effect and between-study variance.}

\item{obj}{character string to specify the objective. "Estimate", which set to default, is overall treatment effect. "Predict" is predictive treatment effect.}

\item{contour_val}{vector of the thresholds for contour plot. By default, the line of (0.01, 0.05, 0.2, 0.4, 0.6, 0.8) is plotted.}

\item{plot}{If TRUE, the contour plot is created.}
}
\value{
overall results for M-estimated random-effects meta-analysis with logarithmic absolute convex function.
}
\description{
Random-effects meta-analysis with M-estimator
}
