% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsim_th_tau2.R
\name{rsim_th_tau2}
\alias{rsim_th_tau2}
\title{Simultaneous distribution of overall treatment effect and between-study variance}
\usage{
rsim_th_tau2(n, theta = 0, tau2, K, vi, method = "DL")
}
\arguments{
\item{n}{number of random variables.}

\item{theta}{value of true overall effect that define 0 in default.}

\item{tau2}{between-study variance.}

\item{K}{number of studies.}

\item{vi}{vector of length K with the corresponding sample variances.}

\item{method}{character string to specify the estimating method of between-study variance. By default (method="DL"), DerSimonian-Laird estimator is obtained. When method="SJ", Sidik-Jonkman estimator is obtained.}
}
\value{
value of simultaneous distribution of overall treatment effect and between-study variance.
}
\description{
Simultaneous distribution of overall treatment effect and between-study variance
}
